backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
SQLAlchemy==2.0.29
sqlalchemy_utils==0.41.2
python-dotenv==1.0.1
pydantic-settings==2.2.1
asyncpg==0.29.0
alembic==1.13.1
